+++ src/components/article_link/ArticleLink.module.scss
@@ -13,7 +13,7 @@
 
 .lists {
   display: block;
-  padding: 12px 0;
+  padding: 16px 0;
   border-bottom: 1px solid $border_color;
 
   @include hover {
