+++ src/components/features/zennRss/constants/index.ts
@@ -1,12 +1,9 @@
-import { MemberType } from '@components/zenn_posts/zennBuilder';
+import { MemberType } from '@components/features/zennRss/zennBuilder';
 
 export const members: MemberType[] = [
   {
     id: 'catnose',
     name: 'CatNose',
-    // role: 'CTO',
-    // bio: 'デザインが好きなプログラマー。開発者向けの情報共有プラットフォームzenn.devを開発しています。',
-    // avatarSrc: '/avatars/catnose.jpg',
     sources: ['https://zenn.dev/tsue/feed'],
     includeUrlRegex: 'zenn.dev',
     // twitterUsername: 'catnose99',
