+++ src/pages/_app.tsx
@@ -10,6 +10,7 @@ import Layout from '@components/layout/Layout';
 
 function MyApp({ Component, pageProps }: AppProps) {
   usePageView();
+
   return (
     <>
       <CommonHead />
